+++ Flutter/tencent_conference_uikit/example/lib/common/routers/names.dart
@@ -1,0 +1,7 @@
+class RouteNames {
+  static const commonLogin = '/common_login';
+  static const commonProfile = '/common_profile';
+  static const conferencePrepare = '/conference_prepare';
+  static const conferenceEnterRoom = '/conference_enter_room';
+  static const conferenceCreateRoom = '/conference_create_room';
+}