+++ Flutter/tencent_conference_uikit/example/lib/pages/common/login/widgets/login_button.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+import '../index.dart';
+
+class LoginButtonWidget extends GetView<LoginController> {
+  const LoginButtonWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: double.infinity,
+      child: ElevatedButton(
+        onPressed: () => controller.login(),
+        child: Text(
+          'login'.tr,
+          style: Get.textTheme.bodyMedium,
+        ),
+      ),
+    );
+  }
+}