+++ Flutter/tencent_conference_uikit/lib/common/widgets/bottom_sheet.dart
@@ -44,7 +44,6 @@ Future<void> showConferenceBottomSheet(Widget bottomSheet,
         },
       );
     },
-    routeSettings: const RouteSettings(name: '/bottom_sheet'),
   );
 }