+++ Flutter/tencent_conference_uikit/lib/common/widgets/dialog.dart
@@ -120,6 +120,5 @@ Future<void> showConferenceDialog({
       ),
     ),
     barrierDismissible: barrierDismissible,
-    routeSettings: const RouteSettings(name: '/dialog'),
   );
 }